+++ src/main/java/com/hamhabocca/dallibocca/report/repository/RallyReportRepository.java
@@ -3,7 +3,7 @@ package com.hamhabocca.dallibocca.report.repository;
 import com.hamhabocca.dallibocca.report.entity.RallyReport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RallyReportRepository extends JpaRepository<RallyReport, Integer> {
+public interface RallyReportRepository extends JpaRepository<RallyReport, Long> {
 
 
 }
